refactor(storage): compose parquet config injectors
